+++ blocks/copy-plugin.js
@@ -1,0 +1,16 @@
+/**
+ * @module copy-plugin
+ */
+
+'use strict'
+
+const CopyWebpackPlugin = require('copy-webpack-plugin')
+
+/**
+ * Configuration for CopyWebpackPlugin
+ */
+module.exports = path => ({
+  plugins: [
+    new CopyWebpackPlugin(path)
+  ]
+})